+++ energyexpertexternal-20220923/alibabacloud_energyexpertexternal20220923/client.py
@@ -4257,8 +4257,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
-        if not UtilClient.is_unset(request.data_type):
-            query['dataType'] = request.data_type
+        if not UtilClient.is_unset(request.analysis_type):
+            query['analysisType'] = request.analysis_type
         if not UtilClient.is_unset(request.file_name):
             query['fileName'] = request.file_name
         if not UtilClient.is_unset(request.file_url):
@@ -4303,8 +4303,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
-        if not UtilClient.is_unset(request.data_type):
-            query['dataType'] = request.data_type
+        if not UtilClient.is_unset(request.analysis_type):
+            query['analysisType'] = request.analysis_type
         if not UtilClient.is_unset(request.file_name):
             query['fileName'] = request.file_name
         if not UtilClient.is_unset(request.file_url):